refactor: ItemList & ListTitle js -> ts로 변환
